fix(#2162): read.rows/partitions contract aligned, capture tests isolated, spec docs aligned (review)
