+++ cqlite-flight/src/lib.rs
@@ -24,11 +24,12 @@ pub mod ticket;
 #[cfg(test)]
 mod testutil;
 
-// Issue #2162: OTel-level assertions (phase histograms, bounded attributes,
-// rpc.rows presence) through the shared `observability-testing` capture harness.
-// Gated behind the feature (which pulls in the SDK in-memory exporters) so the
-// default build never links OTel; the feature-independent `StreamProbe` seam
-// tests in `streaming_tests.rs` carry the always-compiled wiring evidence.
-#[cfg(all(test, feature = "observability-testing"))]
-#[path = "metrics_capture_tests.rs"]
-mod metrics_capture_tests;
+// Issue #2162 OTel-level assertions (phase histograms, bounded attributes,
+// rpc.rows presence via the shared `observability-testing` capture harness) live
+// in `tests/metrics_capture_test.rs` — a SEPARATE integration-test binary/process
+// (roborev, matching the #2163 precedent), not a unit-test module here. The
+// capture harness installs a PROCESS-GLOBAL in-memory meter provider on first
+// use; sharing it with this crate's parallel `cargo test --lib` unit-test binary
+// would risk cross-test metric contamination. The feature-independent
+// `StreamProbe`/`ScanProgress` seam tests in `streaming_tests.rs` carry the
+// always-compiled (feature-off-safe) wiring evidence.
